Processing of messages from scoring-module
The thread that binds to the socket just enqueue the (domain,IP)
tuples in a FIFO queue to be consumed by worker threads. Each worker
thread copy their data to the database every X processed messages
and reinitializes itself.

TODO: trigger end of day events for computing the daily scores.